ENG-221: Moved AutoSizer to correct place in component tree + Updated styling for DragNDropPreview for dark/light mode support
