+++ src/renderer/view/components/datasets/drag-n-drop/DragNDropPreview.css
@@ -1,9 +1,17 @@
-.drag-drop-preview::-webkit-scrollbar {
-	height: 8px;
-	width: 8px;
-	background: #3d3d3d;
+.grid-dark::-webkit-scrollbar {
+	height: 6px;
+	background: var(--chakra-colors-thia-gray-600);
 }
 
-.drag-drop-preview::-webkit-scrollbar-thumb {
-	background: #0a0a0a;
+.grid-dark::-webkit-scrollbar-thumb {
+	background: var(--chakra-colors-thia-gray-950);
+}
+
+.grid-light::-webkit-scrollbar {
+	height: 6px;
+	background: var(--chakra-colors-thia-gray-100);
+}
+
+.grid-light::-webkit-scrollbar-thumb {
+	background: var(--chakra-colors-thia-gray-950);
 }
